+++ Editor/Playa/Renderer/Table/TableContentElement.cs
@@ -228,7 +228,7 @@ namespace SaintsField.Editor.Playa.Renderer.Table
                             .Where(each => each != null)
                             .ToDictionary(each => each.name, each => each.Copy());
                         IEnumerable<SaintsFieldInfoName> saintsFieldWithInfos = SaintsEditor
-                            .HelperGetSaintsFieldWithInfo(serializedPropertyDict,  new []{obj0})
+                            .HelperGetSaintsFieldWithInfo(_fieldWithInfo.SerializedProperty.serializedObject, serializedPropertyDict,  new []{obj0})
                             .Where(SaintsEditor.SaintsFieldInfoShouldDraw)
                             .Select(each => new SaintsFieldInfoName(each, AbsRenderer.GetFriendlyName(each)));
 
@@ -340,7 +340,7 @@ namespace SaintsField.Editor.Playa.Renderer.Table
 
                                 int serCount = 0;
                                 foreach (SaintsFieldWithInfo saintsFieldWithInfo in SaintsEditor
-                                             .HelperGetSaintsFieldWithInfo(targetPropertyDict, new[]{targetPropValue})
+                                             .HelperGetSaintsFieldWithInfo(_fieldWithInfo.SerializedProperty.serializedObject, targetPropertyDict, new[]{targetPropValue})
                                              .Where(saintsFieldWithInfo => memberIds.Contains(saintsFieldWithInfo.MemberId)))
                                 {
                                     allSaintsFieldWithInfos.Add(saintsFieldWithInfo);
@@ -395,7 +395,7 @@ namespace SaintsField.Editor.Playa.Renderer.Table
                     (string error, int index, object value) firstPropValue = Util.GetValue(firstProp, info, parentRefreshed);
 
                     IEnumerable<SaintsFieldWithInfo> firstSaintsFieldWithInfos = SaintsEditor
-                        .HelperGetSaintsFieldWithInfo(firstSerializedPropertyDict, new[]{firstPropValue.value})
+                        .HelperGetSaintsFieldWithInfo(_fieldWithInfo.SerializedProperty.serializedObject, firstSerializedPropertyDict, new[]{firstPropValue.value})
                         .Where(SaintsEditor.SaintsFieldInfoShouldDraw);
 
                     Dictionary<string, List<string>> columnToMemberIds = new Dictionary<string, List<string>>();
@@ -498,7 +498,7 @@ namespace SaintsField.Editor.Playa.Renderer.Table
 
                             int serCount = 0;
                             foreach (SaintsFieldWithInfo saintsFieldWithInfo in SaintsEditor
-                                         .HelperGetSaintsFieldWithInfo(targetSerializedPropertyDict, new[]{targetPropValue.value})
+                                         .HelperGetSaintsFieldWithInfo(_fieldWithInfo.SerializedProperty.serializedObject, targetSerializedPropertyDict, new[]{targetPropValue.value})
                                          .Where(saintsFieldWithInfo => memberIds.Contains(saintsFieldWithInfo.MemberId)))
                             {
                                 // Debug.Log($"get {saintsFieldWithInfo}");