+++ Editor/SaintsEditorUIToolkit.cs
@@ -1,10 +1,8 @@
 #if UNITY_2021_3_OR_NEWER && !SAINTSFIELD_UI_TOOLKIT_DISABLE
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using SaintsField.Editor.HeaderGUI;
 using SaintsField.Editor.Playa;
-using SaintsField.Editor.Playa.Renderer.BaseRenderer;
 using SaintsField.Editor.Utils;
 using SaintsField.Playa;
 using UnityEditor;
@@ -17,10 +15,6 @@ namespace SaintsField.Editor
 {
     public partial class SaintsEditor
     {
-
-        [Obsolete("No longer needed")]
-        protected virtual bool TryFixUIToolkit => false;
-
         private void OnHeaderButtonClickUIToolkit()
         {
             _toolbarSearchField.style.display = _searchableShown ? DisplayStyle.Flex : DisplayStyle.None;
@@ -50,14 +44,14 @@ namespace SaintsField.Editor
             Type objectType = target.GetType();
             IPlayaClassAttribute[] playaClassAttributes = ReflectCache.GetCustomAttributes<IPlayaClassAttribute>(objectType);
 
-            foreach (ISaintsRenderer saintsRenderer in GetClassStructRenderer(objectType, playaClassAttributes, serializedObject, targets))
-            {
-                VisualElement ve = saintsRenderer.CreateVisualElement();
-                if(ve != null)
-                {
-                    root.Add(ve);
-                }
-            }
+            // foreach (ISaintsRenderer saintsRenderer in GetClassStructRenderer(objectType, playaClassAttributes, serializedObject, targets))
+            // {
+            //     VisualElement ve = saintsRenderer.CreateVisualElement();
+            //     if(ve != null)
+            //     {
+            //         root.Add(ve);
+            //     }
+            // }
 
             MonoScript monoScript = GetMonoScript(target);
             if(monoScript)