+++ Runtime/Playa/IPlayaClassAttribute.cs
@@ -1,7 +1,8 @@
+// ReSharper disable once CheckNamespace
 namespace SaintsField.Playa
 {
-    public interface IPlayaClassAttribute
+    public interface IPlayaClassAttribute: IPlayaAttribute
     {
-        
+        bool EndDecorator { get; }
     }
 }