+++ Samples~/Scripts/HeaderComponentExample/WorkTogether.cs
@@ -1,0 +1,11 @@
+using SaintsField.ComponentHeader;
+using SaintsField.Playa;
+
+namespace SaintsField.Samples.Scripts.HeaderComponentExample
+{
+    [Searchable]
+    [HeaderLabel("<color=brown>Header Info")]
+    public class WorkTogether : SaintsMonoBehaviour
+    {
+    }
+}