+++ Samples~/Scripts/SaintsEditor/BelowTextTag/BelowTextChild.cs
@@ -1,0 +1,10 @@
+namespace SaintsField.Samples.Scripts.SaintsEditor.BelowTextTag
+{
+    [Separator("Start Of <color=gray><container.Type />", EAlign.Center)]
+    [AboveText("Start Of <color=gray><container.Type />")]
+    [InfoBox("Start Of <color=gray><container.Type />")]
+    public class BelowTextChild : BelowTextParent
+    {
+        public string child;
+    }
+}