+++ Samples~/Scripts/SaintsEditor/BelowTextTag/ChildSepMono.cs
@@ -1,0 +1,8 @@
+namespace SaintsField.Samples.Scripts.SaintsEditor.BelowTextTag
+{
+    [Separator("Begin of <container.Type>", EAlign.Center)]
+    public class ChildSepMono : AbsSepMono
+    {
+        public string childField;
+    }
+}